Move Command class from spi.common to scheduler
