+++ heron/scheduler-core/src/java/com/twitter/heron/scheduler/Command.java
@@ -1,0 +1,33 @@
+//  Copyright 2017 Twitter. All rights reserved.
+//
+//  Licensed under the Apache License, Version 2.0 (the "License");
+//  you may not use this file except in compliance with the License.
+//  You may obtain a copy of the License at
+//
+//    http://www.apache.org/licenses/LICENSE-2.0
+//
+//  Unless required by applicable law or agreed to in writing, software
+//  distributed under the License is distributed on an "AS IS" BASIS,
+//  WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+//  See the License for the specific language governing permissions and
+//  limitations under the License.
+
+package com.twitter.heron.scheduler;
+
+/***
+ * This enum defines commands invoked from heron client
+ */
+public enum Command {
+  // TODO(mfu): Move ACTIVATE & DEACTIVATE out? They are non-related to Scheduling
+  SUBMIT,
+  KILL,
+  ACTIVATE,
+  DEACTIVATE,
+  UPDATE,
+  RESTART;
+
+  public static Command makeCommand(String commandString) {
+    return Command.valueOf(commandString.toUpperCase());
+  }
+}
+
