+++ heron/scheduler-core/tests/java/com/twitter/heron/scheduler/client/HttpServiceSchedulerClientTest.java
@@ -25,8 +25,8 @@ import org.powermock.core.classloader.annotations.PrepareForTest;
 import org.powermock.modules.junit4.PowerMockRunner;
 
 import com.twitter.heron.proto.scheduler.Scheduler;
+import com.twitter.heron.scheduler.Command;
 import com.twitter.heron.scheduler.utils.SchedulerUtils;
-import com.twitter.heron.spi.common.Command;
 import com.twitter.heron.spi.common.Config;
 import com.twitter.heron.spi.utils.NetworkUtils;
